pyproject.toml
[tool.poetry]
name = "khantoreservations"
version = "0.1.0"
description = ""
authors = ["Cácio Lucas <cclucas060901@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
django-rest-framework = "^0.1.0"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.1"
dj-database-url = "^1.2.0"
python-decouple = "^3.6"
drf-yasg = "^1.21.4"

[tool.poetry.dev-dependencies]
isort = "^5.11.4"
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
